+++ NEWS.md
@@ -2,7 +2,7 @@ Version 1.3.2
 ==============
 
 ### New Features
-* Ordination with predictors (num.RR,num.lv.c) is now implemented with constrained optimization routines (alabama,nloptr) as long as the canonical coefficients are treated as fixed-effects. This follows from the necessary identifability constraints. 
+* Ordination with predictors (num.RR,num.lv.c) is now implemented with constrained optimization routines (alabama,nloptr) as long as the canonical coefficients are treated as fixed-effects. This follows from the necessary identifiability constraints. 
 
 * The reduced-rank approximated predictor slopes of a multivariate regression can now be plotted (with confidence intervals) using coefplot. Not available yet for quadratic effects.
 
@@ -15,6 +15,8 @@ Version 1.3.2
 * An extended version of the spider dataset has been made available
 
 * Added an option to magnify the x-axis labels in coefplot
+
+* Site names present as row labels in the response data are now shown in the ordination plot
 
 
 ### Bug Fixes